refactor(grading): move launch logic into features/grading + i18n
Relocate the non-routing code out of the /app/grading/launch route folder
so the App Router tree holds only routing entries (route/page/layout),
matching the convention used elsewhere in the codebase.

- _lib.ts            -> src/features/grading/launch.ts
- _errors.ts         -> src/features/grading/errors.ts (copy removed)
- select/actions.ts  -> src/features/grading/actions.ts
- select/project-picker.tsx -> src/features/grading/project-picker.tsx
- error COPY         -> src/i18n/en/grading.ts (export const messages)

App Router files now import directly from @/features/grading/* and
@/i18n/en/grading (no barrel). Drop export on four launch.ts symbols only
used internally. Pure relocation; behavior unchanged.
